removed dependencies from kt files, renamed SOC to SOA and added output when lib is not present

+++ src/main/jvm/de/rub/bi/inf/openbimrl/functions/ifc/FilterByElement.kt
@@ -20,7 +20,7 @@ class FilterByElement(nodeProxy: NodeProxy) : NativeFunction(nodeProxy) {
                 setResult(memoryStructure.at, emptyList<IfcPointer>())
             else setResult(
                 memoryStructure.at,
-                memoryStructure.memory.getLongArray(0, memoryStructure.size).map { IfcPointer(it) })
+                memoryStructure.memory.getLongArray(0, memoryStructure.size.toInt()).map { IfcPointer(it) })
         }
         memoryQueue.clear() // technically this should automatically happen.
     }
